// layouts/home-page/styled-skills-props.d.ts
export interface StyledContainerProps {
  display?: string;
  widthVal?: string;
  heightVal?: string;
  maxWidthVal?: string;
  maxHeightVal?: string;
  padding?: string;
  margin?: string;
}
export interface StyledGridContainerProps {
  gridCols?: string;
}
export interface StyledFontProps {
  fontWeight?: string;
  fontSizeMin?: string;
  fontSizeInitial?: string;
  fontSizeMax?: string;
}